Clean up the rpc.serverInfo implementation.
- Remove the RPCServerInfo client helper, it was unused outside tests.
- Remove the methodFunc wrapper type, it is no longer necessary.
- Fix up tests.

Updates #46.
